requirements.txt
django==1.2.3
django-autoslug
django-debug-toolbar
# django-haystack -- installed directly from trunk below
django-markitup
django-taggit
django-taggit-templatetags
docutils
easy_thumbnails
markdown
pep8
PIL
simplejson
south
whoosh

-e git+http://github.com/toastdriven/django-haystack.git#egg=haystack

